Initial setup for the prisma-fmt wasm build
Ported over from the experimental
https://github.com/tomhoule/prisma-fmt-build repo.

This is about building and publishing the wasm version of prisma-fmt.

- The code in this repo will contain no business logic, it is purely a
wrapper that exposes the public API of prisma-fmt from prisma-engines
as a WebAssembly module.
- The GitHub action to build and publish prisma-fmt should succeed or
fail independently from the more important engines actions.

- Why in this repo? Because we want to trigger the build on dispatch,
and prisma-fmt cannot easily be built in the engines pipeline, because
it requires a different toolchain and can't be built on some of the
platforms we support (namely: alpine), but since we only care about
the resulting cross-platform `.wasm` file, this is not an obstacle.

If this experiment works out, other Rust code from prisma-engines can
be published as npm WASM packages in the future in a similar way.

- Why nix? It's not a conceptual and operational dependency to take
lightly, but I've weighed the tradeoffs, and I believe we will deliver
working wasm releases more reliably with it. It is also a
standardized, reproducible way to do the build automation beyond
`cargo` that is necessary to build the final npm package. The
alternatives would be shell scripts or ad-hoc rust build automation
like prisma/engineer, with a separate compilation and release process.

Future work:

- Add a checkPhase to the flake.nix to validate that we produced the
expected JS/TS files and a non-empty WASM module.
- Test the npm package before publishing it.

---

Issue: https://github.com/prisma/language-tools/issues/913