Fix exit code of updateLocks script
Since we use `set -e` to be stricter about failures, and propagate the
exit code of `nix build` by assigning to a _local_ variable first, the
build we expect to fail in order to get the sha of datamodel causes the
whole updateLocks script to return exit code 1.

We fix this by swallowing the exit code in that step.